refactor index files
